+++ packages/cli/src/commands/inspect/index.ts
@@ -76,6 +76,7 @@ export default async function inspect(client: Client) {
   telemetry.trackCliOptionTimeout(parsedArguments.flags['--timeout']);
   telemetry.trackCliFlagLogs(parsedArguments.flags['--logs']);
   telemetry.trackCliFlagWait(parsedArguments.flags['--wait']);
+  telemetry.trackCliFlagJson(parsedArguments.flags['--json']);
 
   // validate the timeout
   const timeout = ms(parsedArguments.flags['--timeout'] ?? '3m');
@@ -103,6 +104,7 @@ export default async function inspect(client: Client) {
   const until = Date.now() + timeout;
   const wait = parsedArguments.flags['--wait'] ?? false;
   const withLogs = parsedArguments.flags['--logs'];
+  const asJson = parsedArguments.flags['--json'] ?? false;
   const startTimestamp = Date.now();
 
   try {
@@ -116,7 +118,7 @@ export default async function inspect(client: Client) {
   let deployment = await getDeployment(client, contextName, deploymentIdOrHost);
 
   let abortController: AbortController | undefined;
-  if (withLogs) {
+  if (withLogs && !asJson) {
     let promise: Promise<void>;
     ({ abortController, promise } = displayBuildLogs(client, deployment, wait));
     if (wait) {
@@ -140,7 +142,11 @@ export default async function inspect(client: Client) {
       break;
     }
   }
-  if (withLogs) {
+
+  if (asJson) {
+    output.stopSpinner();
+    await printJson({ deployment, contextName, client });
+  } else if (withLogs) {
     print(`${chalk.cyan('status')}\t${stateString(deployment.readyState)}\n`);
   } else {
     await printDetails({ deployment, contextName, client, startTimestamp });
@@ -262,6 +268,48 @@ async function printDetails({
     print(indent(routesList(routes), 4));
     print(`\n\n`);
   }
+}
+
+async function printJson({
+  deployment,
+  contextName,
+  client,
+}: {
+  deployment: Deployment;
+  contextName: string | null;
+  client: Client;
+}): Promise<void> {
+  const {
+    id,
+    name,
+    url,
+    createdAt,
+    routes,
+    readyState,
+    alias: aliases,
+    target,
+    customEnvironment,
+  } = deployment;
+
+  const { builds } =
+    deployment.version === 2
+      ? await client.fetch<{ builds: Build[] }>(`/v11/deployments/${id}/builds`)
+      : { builds: [] };
+
+  const jsonOutput = {
+    id,
+    name,
+    url,
+    target: customEnvironment?.slug ?? target ?? 'preview',
+    readyState,
+    createdAt,
+    ...(aliases && aliases.length > 0 && { aliases }),
+    ...(builds.length > 0 && { builds }),
+    ...(Array.isArray(routes) && routes.length > 0 && { routes }),
+    ...(contextName && { contextName }),
+  };
+
+  client.stdout.write(`${JSON.stringify(jsonOutput, null, 2)}\n`);
 }
 
 function exitCode(state: Deployment['readyState']) {